// src/types/schema.ts
export type ID1 = string;
export type SchemaKey1 = "Person";
/**
 * An ORCID (orcid.org) identifier for an individual.
 */
export type AnORCIDIdentifier = string;
/**
 * Use the format: familyname, given names ...
 */
export type Name = string;
export type Email = string;
export type URL = string;
/**
 * An enumeration of roles
 */
export type RoleType =
  | "dcite:Author"
  | "dcite:Conceptualization"
  | "dcite:ContactPerson"
  | "dcite:DataCollector"
  | "dcite:DataCurator"
  | "dcite:DataManager"
  | "dcite:FormalAnalysis"
  | "dcite:FundingAcquisition"
  | "dcite:Investigation"
  | "dcite:Maintainer"
  | "dcite:Methodology"
  | "dcite:Producer"
  | "dcite:ProjectLeader"
  | "dcite:ProjectManager"
  | "dcite:ProjectMember"
  | "dcite:ProjectAdministration"
  | "dcite:Researcher"
  | "dcite:Resources"
  | "dcite:Software"
  | "dcite:Supervision"
  | "dcite:Validation"
  | "dcite:Visualization"
  | "dcite:Funder"
  | "dcite:Sponsor"
  | "dcite:StudyParticipant"
  | "dcite:Affiliation"
  | "dcite:EthicsApproval"
  | "dcite:Other";
/**
 * Role(s) of the contributor. Multiple roles can be selected.
 */
export type Role = RoleType[];
/**
 * A flag to indicate whether a contributor should be included when generating a citation for the item.
 */
export type IncludeContributorInCitation = boolean;
/**
 * Identifier associated with a sponsored or gift award.
 */
export type IdentifierForAnAward = string;
/**
 * Uniform resource identifier
 */
export type ID2 = string;
export type SchemaKey2 = "Affiliation";
/**
 * Use an ror.org identifier for institutions.
 */
export type ARorOrgIdentifier = string;
/**
 * Name of organization
 */
export type Name1 = string;
/**
 * An organization that this person is affiliated with.
 */
export type Affiliation = Affiliation1[];
/**
 * Uniform resource identifier
 */
export type ID3 = string;
export type SchemaKey3 = "Organization";
/**
 * Use an ror.org identifier for institutions.
 */
export type ARorOrgIdentifier1 = string;
export type Name2 = string;
export type Email1 = string;
export type URL1 = string;
/**
 * Role(s) of the contributor. Multiple roles can be selected.
 */
export type Role1 = RoleType[];
/**
 * A flag to indicate whether a contributor should be included when generating a citation for the item
 */
export type IncludeContributorInCitation1 = boolean;
/**
 * Identifier associated with a sponsored or gift award.
 */
export type IdentifierForAnAward1 = string;
/**
 * Uniform resource identifier
 */
export type ID4 = string;
export type SchemaKey4 = "ContactPoint";
/**
 * Email address of contact.
 */
export type Email2 = string;
/**
 * A Web page to find information on how to contact.
 */
export type URL2 = string;
/**
 * Contact for the organization
 */
export type OrganizationContactInformation = ContactPoint[];
// ...
export interface Person {
  id?: ID1;
  schemaKey: SchemaKey1;
  identifier?: AnORCIDIdentifier;
  name: Name;
  email?: Email;
  url?: URL;
  roleName?: Role;
  includeInCitation?: IncludeContributorInCitation;
  awardNumber?: IdentifierForAnAward;
  affiliation?: Affiliation;
  [k: string]: unknown;
}
export interface Affiliation1 {
  id?: ID2;
  schemaKey: SchemaKey2;
  identifier?: ARorOrgIdentifier;
  name: Name1;
  [k: string]: unknown;
}
export interface Organization {
  id?: ID3;
  schemaKey: SchemaKey3;
  identifier?: ARorOrgIdentifier1;
  name?: Name2;
  email?: Email1;
  url?: URL1;
  roleName?: Role1;
  includeInCitation?: IncludeContributorInCitation1;
  awardNumber?: IdentifierForAnAward1;
  contactPoint?: OrganizationContactInformation;
  [k: string]: unknown;
}
export interface ContactPoint {
  id?: ID4;
  schemaKey: SchemaKey4;
  email?: Email2;
  url?: URL2;
  [k: string]: unknown;
}